requirements.txt
python-dotenv
polars
jupyter
pyarrow
pandas
great-tables